// src/utils/types/shared/common.ts
// Define a type for the supported locales
export type Locale = "en" | "es" | "fr" | "ar";

export interface HomeNavigationItems {
  name: string;
  link: string;
}

export interface Sponsors {
  name: string;
  imageUrl: string;
  imageWidth: number;
  imageHeight: number;
  mobileWidth: number;
  mobileHeight: number;
}

export interface KeyAboutItems {
  title: string;
  description: string;
  link: string;
}

export interface HomeNavigationItems {
  name: string;
  link: string;
}